requirements.txt
pandas~=2.2
requests~=2.32
beautifulsoup4~=4.12
pytest-cov~=6.0
